backend/requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
django-stubs==5.1.0
django-stubs-ext==5.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.1
djangorestframework==3.12.4
django-environ==0.7.0
django-filter==21.1
django-autoslug==1.9.8
psycopg2-binary==2.9.1
flake8==3.9.2
black==21.6b0
isort==5.9.3
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
pytest-django==4.4.0
pytest-factoryboy==2.1.0
Faker==8.9.1
pytest-cov==2.12.1
